requirements.txt
pandas 
numpy
scikit-learn
pymongo
certifi
python-dotenv
pymongo[srv]==3.6
#-e .